Bringing in line with featureImpl

+++ src/main/java/com/BJJ/BJJSite/Classes/Administrator.java
@@ -35,7 +35,12 @@ public class Administrator extends Employee {
         }
     }
 
-    public Long getUserId() {
+    /**
+     * Retrieves the user ID of the Administrator.
+     * 
+     * @return The user ID.
+     */
+    public Integer getUserId() {
         return userId;
     }
